+++ hera_sim/__init__.py
@@ -6,7 +6,7 @@ from . import io
 from . import noise
 from . import rfi
 from . import sigchain
-from . import vis
+from hera_sim.visibilities import simulators
 from . import version
 from . import eor
 from . import utils