+++ hera_sim/tests/test_noise.py
@@ -50,17 +50,14 @@ class TestNoise(unittest.TestCase):
         np.random.seed(0)
         nos_jy = noise.sky_noise_jy(tsky, fqs, lsts, inttime=None, omega_p=omp)
         np.testing.assert_allclose(
-            np.std(nos_jy, axis=0) / scaling * np.sqrt(1e6 * aipy.const.sidereal_day/500), 1.0, atol=0.1
+            np.std(nos_jy, axis=0) / scaling * np.sqrt(1e6 * aipy.const.sidereal_day / 500), 1.0,
+            atol=0.1
         )
         np.random.seed(0)
         nos_jy = noise.sky_noise_jy(tsky, fqs, lsts, B=.1, inttime=10.7, omega_p=omp)
         np.testing.assert_allclose(
             np.std(nos_jy, axis=0) / scaling * np.sqrt(1e8 * 10.7), 1.0, atol=0.1
         )
-        # tsky = noise.resample_Tsky(fqs,lsts,noise.HERA_Tsky_mdl['xx'])
-        # nos_jy = noise.sky_noise_jy(tsky, fqs, lsts)
-        # import uvtools, pylab as plt
-        # uvtools.plot.waterfall(nos_jy, mode='real'); plt.show()
 
 
 if __name__ == "__main__":