+++ hera_sim/tests/test_utils.py
@@ -1,3 +1,4 @@
+from __future__ import print_function
 import unittest
 from hera_sim import utils, noise
 from hera_sim.data import DATA_PATH
@@ -65,7 +66,7 @@ class TestUtils(unittest.TestCase):
         nt.assert_almost_equal(np.sum(ff[50]), np.sum(ff[-50]), 41)
 
         ff = utils.gen_fringe_filter(lsts, fqs, bl_len_ns, filter_type='gauss', fr_width=1e-4)
-        nt.assert_almost_equal(np.sum(ff[50]), 63.06179083841268)
+        nt.assert_almost_equal(np.sum(ff[50]), 63.06179070109816)
 
         ff = utils.gen_fringe_filter(lsts, fqs, bl_len_ns, filter_type='custom', FR_filter=fr_filt, FR_frates=fr_frates, FR_freqs=fr_freqs)
         nt.assert_almost_equal(np.sum(ff[50]), 14.66591593210259, places=3)
@@ -122,7 +123,7 @@ def test_delay_filter_norm():
 
     out /= nreal
 
-    print(out, np.sum(tsky**2))
+    print((out, np.sum(tsky**2)))
     assert np.isclose(out, np.sum(tsky**2), atol=0, rtol=1e-2)